Initial setup to publish to rubygems
This commit adds the initial setup to publish this as a ruby gem,
this also moves all of the existing code under a folder called
original, and we will move out the files from there as we progress
through with the development.

This commit also updates the standard gem licensing to NIST's
base license and it also clearly specifies the purpose for this
specific fork.