Restructure express document parsing code
This commit takes the initial step to restructure existing codebase,
and it tries to simplify each of the interface and puts them to their
dedicated class. This also expose only a single interface that accept
an Nokogiri document and then instantiate the instances.

This also introduce Nokogiri to parse the xml document, but it keeps
the code as similar to the original version so we can easily migrate
to this version, but once this is functionality then we should look
into each of these classes and do necessary refactoring.

This also import all of the existing post processing codes, but this
were causing some issue, like slow performance. So, we are disabling
this for now, but we've added some @todo comment in each of those
inherited codes.